Move editor factories to their own submodule.
