+++ traits/trait_handlers.py
@@ -2011,7 +2011,7 @@ class TraitList(TraitHandler):
         return "a list of %s%s" % (size, info)
 
     def get_editor(self, trait):
-        from traits.traits import list_editor
+        from traits.editor_factories import list_editor
 
         return list_editor(trait, self)
